Script is ready to be merged.
